Remove (int) casts before get_option calls in tests

+++ tests/test-clearance-page-exists.php
@@ -46,7 +46,7 @@ class Test_Clearance_Page_Exists extends WP_UnitTestCase {
 
 	public function test_clearance_page_exists_returns_true_when_option_is_numeric_string(): void {
 		// Arrange.
-		$existing_id = (int) get_option( CLEARANCE_PAGE_OPTION );
+		$existing_id = get_option( CLEARANCE_PAGE_OPTION );
 		if ( $existing_id > 0 ) {
 			wp_delete_post( $existing_id, true );
 		}
@@ -64,7 +64,7 @@ class Test_Clearance_Page_Exists extends WP_UnitTestCase {
 
 	public function test_clearance_page_exists_returns_false_when_no_page_exists(): void {
 		// Arrange.
-		$existing_id = (int) get_option( CLEARANCE_PAGE_OPTION );
+		$existing_id = get_option( CLEARANCE_PAGE_OPTION );
 		if ( $existing_id > 0 ) {
 			wp_delete_post( $existing_id, true );
 		}
@@ -79,7 +79,7 @@ class Test_Clearance_Page_Exists extends WP_UnitTestCase {
 
 	public function test_clearance_page_exists_returns_true_when_page_exists(): void {
 		// Arrange.
-		$existing_id = (int) get_option( CLEARANCE_PAGE_OPTION );
+		$existing_id = get_option( CLEARANCE_PAGE_OPTION );
 		if ( $existing_id > 0 ) {
 			wp_delete_post( $existing_id, true );
 		}
@@ -95,13 +95,13 @@ class Test_Clearance_Page_Exists extends WP_UnitTestCase {
 
 	public function test_clearance_page_exists_returns_false_after_page_is_trashed(): void {
 		// Arrange.
-		$existing_id = (int) get_option( CLEARANCE_PAGE_OPTION );
+		$existing_id = get_option( CLEARANCE_PAGE_OPTION );
 		if ( $existing_id > 0 ) {
 			wp_delete_post( $existing_id, true );
 		}
 		delete_option( CLEARANCE_PAGE_OPTION );
 		create_clearance_page();
-		$page_id = (int) get_option( CLEARANCE_PAGE_OPTION );
+		$page_id = get_option( CLEARANCE_PAGE_OPTION );
 		wp_trash_post( $page_id );
 
 		// Act.
@@ -113,13 +113,13 @@ class Test_Clearance_Page_Exists extends WP_UnitTestCase {
 
 	public function test_clearance_page_exists_returns_false_after_page_is_deleted(): void {
 		// Arrange.
-		$existing_id = (int) get_option( CLEARANCE_PAGE_OPTION );
+		$existing_id = get_option( CLEARANCE_PAGE_OPTION );
 		if ( $existing_id > 0 ) {
 			wp_delete_post( $existing_id, true );
 		}
 		delete_option( CLEARANCE_PAGE_OPTION );
 		create_clearance_page();
-		$page_id = (int) get_option( CLEARANCE_PAGE_OPTION );
+		$page_id = get_option( CLEARANCE_PAGE_OPTION );
 		wp_delete_post( $page_id, true );
 
 		// Act.
